backend/requirements.txt
flask
flask-cors
firebase-admin
google-generativeai
Pillow
python-dotenv
beautifulsoup4